loader/addons: move addon loading code into addons module
The loader module now only owns the XML/Lua file-loading engine
(forAddon/bindings). Addon lifecycle management (TOC resolution,
initAddons, loadAddon, loadFrameXml, saveAllVariables) moves into the
addons module, which now depends on loader for file loading instead of
loader depending on addons for addon data.
